refactor: modularize configuration, exceptions, and utilities
